refactor(ui): drop types.ts, import Catalog from types.gen directly
Moves `EndpointAnalytics` and `AnalyticsData` (parsed Prometheus
output) into prometheus.ts where the rest of the Prometheus types
already live. With nothing else left, types.ts is deleted.

Catalog consumers now import from `@/lib/types.gen` and analytics
consumers from `@/lib/prometheus`, each from its real source.
