+++ martin/martin-ui/__tests__/components/catalogs/font.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { FontCatalog } from '@/components/catalogs/font';
-import type { Catalog } from '@/lib/types';
+import type { Catalog } from '@/lib/types.gen';
 
 // Mock the toast hook
 vi.mock('@/hooks/use-toast', () => ({
