+++ martin/martin-ui/src/components/dialogs/sprite-download.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/components/ui/dialog';
 import { useCopyToClipboard } from '@/hooks/use-copy-to-clipboard';
 import { buildMartinUrl } from '@/lib/api';
-import type { Catalog } from '@/lib/types';
+import type { Catalog } from '@/lib/types.gen';
 
 interface SpriteDownloadDialogProps {
   name: string;
